+++ crates/coop/Cargo.toml
@@ -38,7 +38,6 @@ registry = { path = "../registry" }
 settings = { path = "../settings" }
 client_keys = { path = "../client_keys" }
 auto_update = { path = "../auto_update" }
-signer_proxy = { path = "../signer_proxy" }
 
 rust-i18n.workspace = true
 i18n.workspace = true
@@ -60,6 +59,5 @@ smallvec.workspace = true
 smol.workspace = true
 futures.workspace = true
 oneshot.workspace = true
-webbrowser.workspace = true
 
 tracing-subscriber = { version = "0.3.18", features = ["fmt"] }
