+++ crates/identity/Cargo.toml
@@ -10,11 +10,13 @@ global = { path = "../global" }
 common = { path = "../common" }
 client_keys = { path = "../client_keys" }
 settings = { path = "../settings" }
+signer_proxy = { path = "../signer_proxy" }
 
 nostr-sdk.workspace = true
 nostr-connect.workspace = true
-oneshot.workspace = true
+smol.workspace = true
 gpui.workspace = true
 anyhow.workspace = true
 log.workspace = true
 smallvec.workspace = true
+webbrowser.workspace = true
