+++ crates/registry/Cargo.toml
@@ -17,7 +17,6 @@ itertools.workspace = true
 chrono.workspace = true
 smallvec.workspace = true
 smol.workspace = true
-oneshot.workspace = true
 log.workspace = true
 
 fuzzy-matcher = "0.3.7"
